requirements.txt
# API und Netzwerk
aiohttp==3.9.3
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0

# Datenverarbeitung und Analyse
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
scikit-learn==1.4.1
ta-lib==0.4.28

# KI und Machine Learning
openai==1.12.0
google-generativeai==0.3.2
torch==2.2.1
transformers==4.38.2

# Technische Analyse
pandas-ta==0.3.14b0
mplfinance==0.12.10b0
yfinance==0.2.36

# Visualisierung
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Backtesting
backtrader==1.9.78.123
vectorbt==0.25.4

# Monitoring und Logging
prometheus-client==0.20.0
python-telegram-bot==20.8
tenacity==8.2.3

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Entwicklungswerkzeuge
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Dokumentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
nbsphinx==0.9.3 